revisi fitur fasilitas

+++ features/review/dataReview/gorm.go
@@ -6,8 +6,8 @@ import (
 
 type Reviews struct {
 	gorm.Model
-	UserID  uint   `json:"user_id"`
-	RoomID  uint   `json:"room_id"`
-	Rating  int    `json:"rating"`
-	Comment string `json:"comment"`
+	UserID  uint    `json:"user_id"`
+	RoomID  uint    `json:"room_id"`
+	Rating  float32 `json:"rating"`
+	Comment string  `json:"comment"`
 }
